server/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
psycopg[binary]==3.2.12
psycopg-pool==3.2.7
mangum==0.19.0
python-dotenv==1.0.1
